Fixed torset issue
The torset cannot be found with just the daemonset.  The utility,
`sharp_cmd`, needs to be run with all the HCA GUIDs at once.  The node
labeler has been updated to add annotations for the GUIDs and a separate
job has been added to label the torsets, taking the GUIDs from the
annotations.